+++ Mods ChoGGi/Applicants Age/metadata.lua
@@ -1,0 +1,21 @@
+return PlaceObj("ModDef", {
+	"title", "Applicants Age",
+	"id", "ChoGGi_ApplicantsAge",
+	"steam_id", "2427034032",
+	"pops_any_uuid", "94298e32-b112-4fbb-98c0-47005af1f4b6",
+	"lua_revision", 1001514,
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[Applicants will age while waiting for their chance to set foot on the red planet.
+
+
+Requested by bloodnok.]],
+})