+++ Mods ChoGGi/Farm Cheats/metadata.lua
@@ -1,0 +1,23 @@
+return PlaceObj("ModDef", {
+	"title", "Farm Cheats",
+	"id", "ChoGGi_FarmCheats",
+	"lua_revision", 1001514,
+	"steam_id", "2426931293",
+	"pops_any_uuid", "204e9e89-25af-44e3-a6f2-8b9c824d92e5",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagCrops", true,
+	"description", [[Adds button to quickly finish growing current crop, as well as mod options to improve failure/yields.
+
+Mod Options:
+Crops Never Fail: Crops will never fail no matter the conditions (you'll get a random yield amount instead of failing).
+Constant Soil Quality: Soil quality will always be this amount (0 to disable).
+]],
+})