+++ Mods ChoGGi/Fix Tourism Update Colonist Issues/metadata.lua
@@ -4,20 +4,18 @@ return PlaceObj("ModDef", {
 	"lua_revision", 1001514,
 	"steam_id", "2426008475",
 	"pops_any_uuid", "729adc51-a0f3-4948-b89a-f7ee0e8d2266",
-	"version", 4,
+	"version", 5,
 	"version_major", 0,
-	"version_minor", 4,
+	"version_minor", 5,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
-	"has_options", true,
+--~ 	"has_options", true,
 	"TagGameplay", true,
 	"TagOther", true,
 	"description", [[A fix for the colonist suicides introduced in Tito/Tourism update.
 As well as other colonist issues (aging, moving to other domes, probably more).
-
-Includes mod option to disable once official fix is out (so no missing mod msg when loading game).
 ]],
 })