+++ Mods ChoGGi/Show Saved Colonies/metadata.lua
@@ -1,17 +1,24 @@
 return PlaceObj("ModDef", {
 	"title", "Show Saved Colonies",
 	"id", "ChoGGi_ShowLastColonies",
-	"lua_revision", 1001514,
-	"version", 12,
-	"version_major", 1,
-	"version_minor", 2,
-	"image", "Preview.png",
 	"steam_id", "1506312571",
 	"pops_any_uuid", "6e82eaa9-3834-4d37-b9f5-51b553c99e47",
+	"lua_revision", 1001514,
+	"version", 13,
+	"version_major", 1,
+	"version_minor", 3,
+	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
-	"description", [[Puts a dome marker on areas you've already been to.
-This also adds the names of the saves to the pointer text.]],
+	"has_options", true,
+	"TagInterface", true,
+	"description", [[Show a dome marker for areas with a saved game, also shows the names of the saves next to the marker.
+
+Mod Options:
+Amount Of Saves: Limit visible amount of saved games (default 5).
+
+Thanks to [url=https://steamcommunity.com/workshop/filedetails/?id=1418440299]SkiRich[/url] for the marker icon.
+]],
 })